update admin UI with primeng

+++ front-end-angular/src/app/components/admin/course/course-instructor-list/course-instructor-list.component.ts
@@ -1,4 +1,5 @@
 import { Component, Input, OnChanges, OnInit, SimpleChanges } from '@angular/core';
+import { ConfirmationService, MessageService } from 'primeng/api';
 import { Instructor } from 'src/app/models/instructor.model';
 import { AdminCourseService } from 'src/app/services/admin/course/course.service';
 
@@ -10,8 +11,11 @@ import { AdminCourseService } from 'src/app/services/admin/course/course.service
 export class AdminCourseInstructorListComponent implements OnChanges {
 
   @Input() courseId?: number;
+  @Input() instructorDialog: boolean;
   instructors: Instructor[];
-  constructor(private courseService: AdminCourseService) { }
+  constructor(private courseService: AdminCourseService,
+    private messageService: MessageService,
+    private confirmationService: ConfirmationService) { }
 
 
   ngOnChanges(): void {
@@ -32,11 +36,21 @@ export class AdminCourseInstructorListComponent implements OnChanges {
   }
 
   deleteInstructorFromCourse(instructorId?: number) {
-    this.courseService.deleteInstructorFromCourse(instructorId, this.courseId).subscribe(
-      () => {
-        this.getAllInstructorForCourseRegistered();
+    this.confirmationService.confirm({
+      message: 'Are you sure you want to delete ?',
+      header: 'Delete Instructor',
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        this.courseService.deleteInstructorFromCourse(instructorId, this.courseId).subscribe(
+          () => {
+            this.getAllInstructorForCourseRegistered();
+          }
+        )
+        this.messageService.add({ severity: 'success', summary: 'Successful', detail: 'Instructor deleted from Course', life: 3000 });
       }
-    )
+    });
+
+   
 
   }
