+++ front-end-angular/src/app/components/admin/student/student-list/student-list.component.ts
@@ -1,5 +1,7 @@
 import { ThisReceiver } from '@angular/compiler';
 import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
+import { ConfirmationService, MessageService } from 'primeng/api';
 
 import { Student } from 'src/app/models/student.model';
 import { AdminStudentService } from 'src/app/services/admin/student/student.service';
@@ -12,6 +14,7 @@ import { AdminStudentService } from 'src/app/services/admin/student/student.serv
 export class AdminStudentListComponent implements OnInit {
 
   students: Student[];
+  studentDialog: boolean;
 
   isAdding = false;
   isEditing = false;
@@ -19,7 +22,10 @@ export class AdminStudentListComponent implements OnInit {
 
   //data binding
   student = new Student;
-  constructor(private studentService: AdminStudentService) { }
+  constructor(private studentService: AdminStudentService,
+    private messageService: MessageService,
+    private confirmationService: ConfirmationService,
+    private router: Router) { }
 
   ngOnInit(): void {
     this.listStudents();
@@ -33,17 +39,47 @@ export class AdminStudentListComponent implements OnInit {
     )
   }
 
-  deleteAStudent(student: Student) {
+  addingStudent() {
+    this.isAdding = true;
     this.isEditing = false;
-    this.isAdding = false;
-    this.studentService.deleteAStudent(student.id).subscribe(
-      () => {
-        this.updateDeletedStudent(student)
-      }
-    )
+    this.studentDialog = true;
   }
 
-  updateAddedStudents(student: Student) {
+  editingStudent(student: Student) {
+    this.isEditing = true;
+    this.isAdding = false;
+    this.student = { ...student }
+    this.studentDialog = true;
+
+  }
+
+  deleteStudent(student: Student) {
+    this.confirmationService.confirm({
+      message: 'Are you sure you want to delete ' + student.firstName + ' ' + student.lastName + '?',
+      header: 'Confirm',
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        this.studentService.deleteAStudent(student.id).subscribe(
+          () => {
+            this.updateDeletedStudent(student)
+          }
+        )
+        this.messageService.add({ severity: 'success', summary: 'Successful', detail: 'Student Deleted', life: 3000 });
+      }
+    });
+  }
+
+  // deleteAStudent(student: Student) {
+  //   this.isEditing = false;
+  //   this.isAdding = false;
+  //   this.studentService.deleteAStudent(student.id).subscribe(
+  //     () => {
+  //       this.updateDeletedStudent(student)
+  //     }
+  //   )
+  // }
+
+  updateAddedStudents() {
     this.listStudents();
   }
 
@@ -78,6 +114,22 @@ export class AdminStudentListComponent implements OnInit {
   adding() {
     this.isAdding = !this.isAdding;
     this.isEditing = false;
+  }
+
+  logout() {
+    this.confirmationService.confirm({
+      message: 'Are you sure you want to log out?',
+      header: 'Confirm',
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        localStorage.removeItem("userId");
+        localStorage.removeItem("userType");
+        this.router.navigate(['']);
+      }
+    });
+
+
+
   }
